+++ src/main/java/frc/robot/autonomous/AutonUpperScoring.java
@@ -2,8 +2,8 @@ package frc.robot.autonomous;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.commandgroups.UpperScoring;
-import frc.robot.subsystems.leds.PurpleLed;
-import frc.robot.subsystems.leds.YellowLed;
+import frc.robot.subsystems.leds.command.PurpleLed;
+import frc.robot.subsystems.leds.command.YellowLed;
 
 public class AutonUpperScoring extends SequentialCommandGroup {
