+++ src/main/java/frc/robot/subsystems/leds/command/Blink.java
@@ -1,0 +1,2 @@
+package frc.robot.subsystems.leds.command;public class Blink {
+}
