+++ src/main/java/frc/robot/subsystems/leds/command/PurpleLed.java
@@ -1,6 +1,7 @@
-package frc.robot.subsystems.leds;
+package frc.robot.subsystems.leds.command;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
+import frc.robot.subsystems.leds.Leds;
 
 public class PurpleLed extends InstantCommand {
